separates core and file_maangement into static libraries
