+++ CMakeLists.txt
@@ -17,8 +17,7 @@ target_link_libraries(App
 if (CMAKE_BUILD_TYPE STREQUAL "Debug")
     # see section Version of CRT on https://ledas.com/post/861-fast-debug-in-visual-c/
     set_target_properties(App PROPERTIES
-        MSVC_RUNTIME_LIBRARY "MultiThreadedDLL"
-        )
+        MSVC_RUNTIME_LIBRARY "MultiThreadedDLL")
 
     # Not sure if this has any affects (can't see the generated assembly currently).
     # However, this flag is needed to generate same assembly from std::algorithms and bare for loops.
