+++ include/Lib/asset_kitchen/gltf/core.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <string>
-
 #include "Lib/core/core.hpp"
 
 namespace GLTF
