+++ include/Lib/core/.pch.hpp
@@ -33,6 +33,7 @@
 #include <variant>
 
 // std containers
+#include <string>
 #include <vector>
 #include <array>
 #include <unordered_map>
