+++ include/Lib/core/core.cpp
@@ -1,0 +1,3 @@
+#pragma message("----Read CORE/core.Cpp----")
+
+#include "core.hpp"
