+++ include/Lib/file_management/core.hpp
@@ -1,49 +1,23 @@
 #pragma once
+#pragma message("----Read FILE/core.Hpp----")
 
 #include ".pch.hpp"
 
-// This does not play well with precompiled headers, so it will be included in a regular header
-#define STB_IMAGE_IMPLEMENTATION
-#include <stb_image.h>
+#include "Lib/core/core.hpp"
 
 namespace File
 {
-	ByteBuffer LoadAsBytes(std::filesystem::path const & path)
+	ByteBuffer LoadAsBytes(std::filesystem::path const & path);
+
+	ByteBuffer LoadAsBytes(std::filesystem::path const & path, usize file_size);
+
+	std::string LoadAsString(std::filesystem::path const & path);
+
+	struct Image
 	{
-		assert(std::filesystem::exists(path));
-		std::basic_ifstream<byte> file(path, std::ios::in | std::ios::binary | std::ios::ate);
-
-		usize file_size = file.tellg();
-		ByteBuffer buffer(file_size);
-
-		file.seekg(0);
-		file.read(buffer.data.get(), file_size);
-
-		return buffer;
-	}
-
-	ByteBuffer LoadAsBytes(std::filesystem::path const & path, usize file_size)
-	{
-		assert(std::filesystem::exists(path));
-		std::basic_ifstream<byte> file(path, std::ios::in | std::ios::binary);
-
-		ByteBuffer buffer(file_size);
-		file.read(buffer.data.get(), file_size);
-
-		return buffer;
-	}
-
-	std::string LoadAsString(std::filesystem::path const & path)
-	{
-		assert(std::filesystem::exists(path));
-		std::basic_ifstream<char> file(path, std::ios::in | std::ios::binary | std::ios::ate);
-
-		usize file_size = file.tellg();
-		std::string buffer(file_size, '\0');
-
-		file.seekg(0);
-		file.read(buffer.data(), file_size);
-
-		return buffer;
-	}
+		ByteBuffer buffer;
+		i32x2 dimensions;
+		i32 channels;
+	};
+	Image LoadImage(std::filesystem::path const & path);
 }
