+++ include/Lib/opengl/.pch.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <regex>
+#include <charconv>
+
 #include <glbinding/gl45core/gl.h>
 #include <glbinding/gl45ext/gl.h>
 #include <glbinding/glbinding.h>
